refactor(utils): introduce core module
