+++ apps/campfire/src/components/Deck/Slide/SlideLayer/index.tsx
@@ -1,6 +1,6 @@
 import { type ComponentChildren, type ComponentType, type JSX } from 'preact'
 import { Layer, type LayerProps } from '../Layer'
-import parseInlineStyle from '@campfire/utils/parseInlineStyle'
+import { parseInlineStyle } from '@campfire/utils/core'
 
 export interface SlideLayerProps
   extends Omit<LayerProps, 'children' | 'className'> {
