+++ apps/campfire/src/hooks/useSerializedDirectiveRunner.ts
@@ -2,7 +2,7 @@ import { useCallback, useMemo } from 'preact/hooks'
 import type { RootContent } from 'mdast'
 import type { ContainerDirective } from 'mdast-util-directive'
 import rfdc from 'rfdc'
-import { evalExpression } from '@campfire/utils/evalExpression'
+import { evalExpression } from '@campfire/utils/core'
 import { runDirectiveBlock } from '@campfire/utils/directiveHelpers'
 import { useDirectiveHandlers } from '@campfire/hooks/useDirectiveHandlers'
 import { useGameStore } from '@campfire/state/useGameStore'
