--- apps/campfire/src/utils/quote.ts
@@ -1,2 +1,0 @@
-/** Pattern matching a string enclosed in matching quotes or backticks. */
-export const QUOTE_PATTERN = /^(['"`])(.*)\1$/
